+++ lib/iex/resources/symbols.rb
@@ -3,6 +3,10 @@ module IEX
     class Symbols < Resource
       property 'symbol'
       property 'exchange'
+      property 'exchange_suffix', from: 'exchangeSuffix'
+      property 'exchange_name', from: 'exchangeName'
+      property 'exchange_segment', from: 'exchangeSegment'
+      property 'exchange_segment_name', from: 'exchangeSegmentName'
       property 'name'
       property 'date', transform_with: ->(v) { Date.parse(v) }
       property 'enabled', from: 'isEnabled'
@@ -12,6 +16,7 @@ module IEX
       property 'iex_id', from: 'iexId'
       property 'figi'
       property 'cik'
+      property 'lei'
 
       alias :enabled? enabled
     end